Seth, Refactored GUI files to be simpler to navigate. Fully functional commit.
