+++ src/gui_tabs/amplifierTab.py
